x64 builds and compiles.   FIX ME: LoadLibrary fails when loading renderer dlls and error code is blank

+++ src/xrCore/FS_impl.h
@@ -158,7 +158,7 @@ IC u32 IReaderBase<T>::find_chunk (u32 ID, BOOL* bCompressed)
 #endif
     if (bCompressed) *bCompressed = dwType & CFS_CompressMark;
 
-    const int dwPos = impl().tell();
+    const u32 dwPos =(u32) impl().tell();
     if (dwPos + dwSize < (u32)impl().length())
     {
         m_last_pos = dwPos + dwSize;
